use repository address as the link to the project
