+++ labtool2.0/src/components/pages/ReviewStudent.js
@@ -13,6 +13,7 @@ import { trimDate } from '../../util/format'
 import { usePersistedState } from '../../hooks/persistedState'
 
 import { FormMarkdownTextArea } from '../MarkdownTextArea'
+import RepoLink from '../RepoLink'
 
 const PreviousWeekDetails = ({ weekData }) => {
   if (!weekData) {
@@ -231,10 +232,7 @@ export const ReviewStudent = props => {
           {' '}
           {studentData.User.firsts} {studentData.User.lastname}{' '}
           <div style={{ display: 'inline-block', padding: '0px 0px 0px 25px' }}>
-            <a href={studentData.github} target="_blank" rel="noopener noreferrer">
-              <Icon name="github" color="black" />
-              {studentData.projectName}
-            </a>
+            {studentData.projectName}: <RepoLink url={studentData.github} />
           </div>
           {studentData.Tags.map(tag => (
             <div key={tag.id}>
